+++ noxfiles/ci_nox.py
@@ -1,23 +1,25 @@
 """Contains the nox sessions used during CI checks."""
+from functools import partial
+from os import environ
+from typing import Callable, Dict
+
 import nox
 
 from constants_nox import (
-    CI_ARGS,
-    COMPOSE_FILE,
-    COMPOSE_SERVICE_NAME,
     IMAGE_NAME,
-    INTEGRATION_COMPOSE_FILE,
     RUN,
+    RUN_COVERAGE,
     RUN_NO_DEPS,
     START_APP,
-    START_APP_WITH_EXTERNAL_POSTGRES,
     WITH_TEST_CONFIG,
 )
-from run_infrastructure import OPS_TEST_DIR, run_infrastructure
+from test_setup_nox import pytest_ctl, pytest_lib, pytest_ops
 from utils_nox import install_requirements
 
 
-# Static Checks
+###################
+## Static Checks ##
+###################
 @nox.session()
 def static_checks(session: nox.Session) -> None:
     """Run the static checks only."""
@@ -32,8 +34,8 @@ def static_checks(session: nox.Session) -> None:
 @nox.parametrize(
     "mode",
     [
-        nox.param("check", id="check"),
         nox.param("fix", id="fix"),
+        nox.param("check", id="check"),
     ],
 )
 def black(session: nox.Session, mode: str) -> None:
@@ -49,8 +51,8 @@ def black(session: nox.Session, mode: str) -> None:
 @nox.parametrize(
     "mode",
     [
-        nox.param("check", id="check"),
         nox.param("fix", id="fix"),
+        nox.param("check", id="check"),
     ],
 )
 def isort(session: nox.Session, mode: str) -> None:
@@ -97,7 +99,9 @@ def xenon(session: nox.Session) -> None:
     session.run(*command)
 
 
-# Fides Checks
+##################
+## Fides Checks ##
+##################
 @nox.session()
 def check_install(session: nox.Session) -> None:
     """Check that fides installs works correctly."""
@@ -162,10 +166,53 @@ def minimal_config_startup(session: nox.Session) -> None:
     session.run(*start_command, external=True)
 
 
-# Pytest
+############
+## Pytest ##
+############
+TEST_GROUPS = [
+    nox.param("ctl-unit", id="ctl-unit"),
+    nox.param("ctl-not-external", id="ctl-not-external"),
+    nox.param("ctl-integration", id="ctl-integration"),
+    nox.param("ctl-external", id="ctl-external"),
+    nox.param("ops-unit", id="ops-unit"),
+    nox.param("ops-integration", id="ops-integration"),
+    nox.param("ops-external-datastores", id="ops-external-datastores"),
+    nox.param("ops-saas", id="ops-saas"),
+    nox.param("lib", id="lib"),
+]
+
+TEST_MATRIX: Dict[str, Callable] = {
+    "ctl-unit": partial(pytest_ctl, mark="unit"),
+    "ctl-not-external": partial(pytest_ctl, mark="not external"),
+    "ctl-integration": partial(pytest_ctl, mark="integration"),
+    "ctl-external": partial(pytest_ctl, mark="external"),
+    "ops-unit": partial(pytest_ops, mark="unit"),
+    "ops-integration": partial(pytest_ops, mark="integration"),
+    "ops-external-datastores": partial(pytest_ops, mark="external_datastores"),
+    "ops-saas": partial(pytest_ops, mark="saas"),
+    "lib": pytest_lib,
+}
+TEST_REPORTER_URL = (
+    "https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64"
+)
+
+
+def validate_test_matrix(session: nox.Session) -> None:
+    """
+    Validates that all test groups are represented in the test matrix.
+    """
+    test_group_ids = sorted([str(param) for param in TEST_GROUPS])
+    test_matrix_keys = sorted(TEST_MATRIX.keys())
+
+    if not test_group_ids == test_matrix_keys:
+        session.error("TEST_GROUPS and TEST_MATRIX do not match")
+
+
 @nox.session()
 def collect_tests(session: nox.Session) -> None:
-    """Run the 'pylint' code linter."""
+    """
+    Collect all pytests as a validity check.
+    """
     session.install(".")
     install_requirements(session)
     command = ("pytest", "tests/", "--collect-only")
@@ -173,171 +220,102 @@ def collect_tests(session: nox.Session) -> None:
 
 
 @nox.session()
-def pytest_lib(session: nox.Session) -> None:
-    """Runs ctl tests."""
+@nox.parametrize(
+    "test_group",
+    TEST_GROUPS,
+)
+def pytest(session: nox.Session, test_group: str) -> None:
+    """
+    Runs Pytests.
+
+    As new TEST_GROUPS are added, the TEST_MATRIX must also be updated.
+    """
     session.notify("teardown")
-    session.run(*START_APP, external=True)
-    run_command = (
+
+    validate_test_matrix(session)
+    coverage_arg = f"--cov-report=lcov:coverage/{test_group}.lcov"
+    TEST_MATRIX[test_group](session=session, coverage_arg=coverage_arg)
+
+
+@nox.session()
+@nox.parametrize(
+    "test_group",
+    TEST_GROUPS,
+)
+def format_coverage(session: nox.Session, test_group: str) -> None:
+    """
+    Generate a Code Climate coverage file from a pytest coverage file.
+
+    Expects a 'coverage/<test_group>.lcov' file to exist.
+    Writes a new file to 'coverage/<test_group>/codeclimate.json'
+    """
+
+    # Run the script that fixes the path
+    session.run(
         *RUN_NO_DEPS,
-        "pytest",
-        "tests/lib/",
+        "python",
+        "scripts/fix_coverage.py",
+        test_group,
+        external=True,
     )
-    session.run(*run_command, external=True)
+
+    # Run the Code Climate Reporter tool
+    download_cc_reporter = f"curl -L {TEST_REPORTER_URL} > ./cc-test-reporter && chmod +x ./cc-test-reporter && "
+    format_test_coverage = f"./cc-test-reporter -d format-coverage -t lcov -o coverage/{test_group}/codeclimate.json coverage/{test_group}-fixed.lcov ;"
+    session.run(
+        *RUN_COVERAGE,
+        "bash",
+        "-c",
+        download_cc_reporter + format_test_coverage,
+        external=True,
+    )
 
 
 @nox.session()
-@nox.parametrize(
-    "mark",
-    [
-        nox.param("unit", id="unit"),
-        nox.param("integration", id="integration"),
-        nox.param("not external", id="not-external"),
-        nox.param("external", id="external"),
-    ],
-)
-def pytest_ctl(session: nox.Session, mark: str) -> None:
-    """Runs ctl tests."""
-    session.notify("teardown")
-    if mark == "external":
-        start_command = (
-            "docker",
-            "compose",
-            "-f",
-            COMPOSE_FILE,
-            "-f",
-            INTEGRATION_COMPOSE_FILE,
-            "up",
-            "-d",
-            IMAGE_NAME,
-        )
-        session.run(*start_command, external=True)
-        run_command = (
-            "docker",
-            "compose",
-            "run",
-            "-e",
-            "SNOWFLAKE_FIDESCTL_PASSWORD",
-            "-e",
-            "REDSHIFT_FIDESCTL_PASSWORD",
-            "-e",
-            "AWS_ACCESS_KEY_ID",
-            "-e",
-            "AWS_SECRET_ACCESS_KEY",
-            "-e",
-            "AWS_DEFAULT_REGION",
-            "-e",
-            "OKTA_CLIENT_TOKEN",
-            "-e",
-            "BIGQUERY_CONFIG",
-            "--rm",
-            CI_ARGS,
-            IMAGE_NAME,
-            "pytest",
-            "-m",
-            "external",
-            "tests/ctl",
-        )
-        session.run(*run_command, external=True)
-    else:
-        session.run(*START_APP, external=True)
-        run_command = (
-            *RUN_NO_DEPS,
-            "pytest",
-            "tests/ctl/",
-            "-m",
-            mark,
-        )
-        session.run(*run_command, external=True)
+def sum_coverage(session: nox.Session) -> None:
+    """
+    Combine multiple Code Climate coverage files.
+
+    4 files are expected for summation, one from each of the safe test groups run in CI.
+    """
+    download_cc_reporter = f"curl -L {TEST_REPORTER_URL} > ./cc-test-reporter && chmod +x ./cc-test-reporter && "
+    sum_test_coverage = (
+        "./cc-test-reporter sum-coverage --parts 4 coverage/**/codeclimate.json ;"
+    )
+
+    session.run(
+        *RUN_NO_DEPS,
+        "bash",
+        "-c",
+        download_cc_reporter + sum_test_coverage,
+        external=True,
+    )
 
 
 @nox.session()
-@nox.parametrize(
-    "mark",
-    [
-        nox.param("unit", id="unit"),
-        nox.param("integration", id="integration"),
-        nox.param("external_datastores", id="external-datastores"),
-        nox.param("saas", id="saas"),
-    ],
-)
-def pytest_ops(session: nox.Session, mark: str) -> None:
-    """Runs fidesops tests."""
-    session.notify("teardown")
-    if mark == "unit":
-        session.run(*START_APP, external=True)
-        run_command = (
-            *RUN_NO_DEPS,
-            "pytest",
-            OPS_TEST_DIR,
-            "-m",
-            "not integration and not integration_external and not integration_saas",
-        )
-        session.run(*run_command, external=True)
-    elif mark == "integration":
-        run_infrastructure(
-            run_tests=True,
-            analytics_opt_out=True,
-            datastores=[],
-            pytest_path=OPS_TEST_DIR,
-        )
-    elif mark == "external_datastores":
-        session.run(*START_APP, external=True)
-        run_command = (
-            "docker",
-            "compose",
-            "run",
-            "-e",
-            "ANALYTICS_OPT_OUT",
-            "-e",
-            "REDSHIFT_TEST_URI",
-            "-e",
-            "SNOWFLAKE_TEST_URI",
-            "-e",
-            "REDSHIFT_TEST_DB_SCHEMA",
-            "-e",
-            "BIGQUERY_KEYFILE_CREDS",
-            "-e",
-            "BIGQUERY_DATASET",
-            "--rm",
-            CI_ARGS,
-            COMPOSE_SERVICE_NAME,
-            "pytest",
-            OPS_TEST_DIR,
-            "-m",
-            "integration_external",
-        )
-        session.run(*run_command, external=True)
-    elif mark == "saas":
-        # This test runs an additional integration Postgres database.
-        # Some connectors cannot be traversed with the standard email
-        # identity and require another dataset to provide a starting value.
-        #
-        #         ┌────────┐                 ┌────────┐
-        # email──►│postgres├──►delivery_id──►│doordash│
-        #         └────────┘                 └────────┘
-        #
-        session.run(*START_APP_WITH_EXTERNAL_POSTGRES, external=True)
-        run_command = (
-            "docker",
-            "compose",
-            "run",
-            "-e",
-            "ANALYTICS_OPT_OUT",
-            "-e",
-            "VAULT_ADDR",
-            "-e",
-            "VAULT_NAMESPACE",
-            "-e",
-            "VAULT_TOKEN",
-            "--rm",
-            CI_ARGS,
-            COMPOSE_SERVICE_NAME,
-            "pytest",
-            OPS_TEST_DIR,
-            "-m",
-            "integration_saas",
-        )
-        session.run(*run_command, external=True)
+def upload_coverage(session: nox.Session) -> None:
+    """
+    Upload a test coverage file.
+
+    This should only be used in CI to keep coverage statistics consistent.
+    """
+    try:
+        cc_reported_key = environ["CC_TEST_REPORTER_ID"]
+    except KeyError:
+        session.error("The CC_TEST_REPORTER_ID env var is required for this session")
+
+    download_cc_reporter = f"curl -L {TEST_REPORTER_URL} > ./cc-test-reporter && chmod +x ./cc-test-reporter && "
+    upload_test_coverage = (
+        f"./cc-test-reporter -d upload-coverage -r {cc_reported_key} ;"
+    )
+
+    session.run(
+        *RUN_NO_DEPS,
+        "bash",
+        "-c",
+        download_cc_reporter + upload_test_coverage,
+        external=True,
+    )
 
 
 @nox.session()