+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The types of changes are:
 ### Added
 
 * Added the `user` command group to the CLI. [#2153](https://github.com/ethyca/fides/pull/2153)
+* Added `Code Climate` test coverage uploads.  [#2198](https://github.com/ethyca/fides/pull/2198)
 * Added the connection key to the execution log [#2100](https://github.com/ethyca/fides/pull/2100)
 * Added endpoints to retrieve DSR `Rule`s and `Rule Target`s [#2116](https://github.com/ethyca/fides/pull/2116)
 * Added Fides version number to account dropdown in the UI [#2140](https://github.com/ethyca/fides/pull/2140)